Move Gemini analysis behind backend API
